Initial commit for SleeperPy library
This push includes the beginnings of a library and program intended
to access the sleeper bot API for fantasy football. As it stands
now, you can use this library to generate a League object, and then
iterate through various teams within that league to look at summary
statistics about their players.

The player data is only downloaded once as currently written, and
code needs to be added in the future to refresh the player data
periodically to reflect the ever-changing NFL landscape