Add support for rotating the player data file once/day
Previously, once the player data file was created we would never
go out and "refresh" the data stored there. With this change, we
will go out and refresh the data in the file once per day (as
per the guidelines set out in the Sleeper API docs). This allows
us to pick up the most up-to-date information without slamming
the network or their endpoints for data.