Add support for unowned teams when parsing league
If a team has been abandoned, when fetching the league the program
would crash on a KeyError because the user was not in our pre-made
lookup program. To guard against that, when we are creating the Team
objects, if we don't have an owner we just add a default "unknown"
owner for that team.