Amends player handling for DEF players
Despite the player data files containing entries for coaches,
apparently it doesn't contain anything for defensive units. Not
only that, but when a team has a DEF on their roster the player key
is the alphabetic abbreviation for that team (JAX) instead of a
player key.

This adds support for a dummy "DEF" player type and position to be
added as a valid player when a Player is fetched by `getPlayer`. Age
is non-applicable because it describes more than one player, the name
is their given player id, and the position is "DEF"